+++ Changes.md
@@ -1,10 +1,13 @@
 * #657 - Allow the import database process to do partial imports
+* #679 - when loading from a database, select the tournament that was in the database
+* #676 - keep tomcat logs from testing out of the distribution
+* #673 - Pass full URL, including query string, when redirected to the login form
 
 Release 13.2.0
 ==============
 
 * #494 - add application icon to get rid of the default Java icon
-* #669 Keep the launcher from running twice
+* #669 - Keep the launcher from running twice
 * #671 - Ensure that the link to the SSL web page uses the host name that the user visited 
 
 Release 13.1.1