+++ Changes.md
@@ -1,3 +1,7 @@
+Release 13.3.0
+==============
+
+* #173 - add tool to edit challenge descriptions
 * #688 - allow the user to specify the filename suffix on subjective score sheets
 * #670 - streamline the export/import process for the 2 server model
 * #690 - remove the ranking report